requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-dotenv
google-genai
protobuf==4.25.3   
anthropic